contrib/starrocks-python-client/requirements-dev.txt
sqlalchemy==2.0.25
sqlalchemy-utils==0.41.2
alembic==1.14.1
pymysql==1.1.1
lark-parser>=0.12.0
pytest>=8.0
sqlacodegen>=3.0